add 3 controllers ‘SessionsController’, ‘UsersController’, ‘StaticPagesController’,add Models folders on the app, move the User.php file to Models folders, add 3 routes, about home, signUp, login
