+++ app/Http/Controllers/SessionsController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class SessionsController extends Controller
+{
+    // 用户登录操作
+    public function create()
+    {
+        return '登录表单';
+    }
+}
