+++ app/Http/Controllers/StaticPagesController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class StaticPagesController extends Controller
+{
+    // 首页展示页面
+    public function home()
+    {
+        return '静态页面';
+    }
+}
