+++ app/Http/Controllers/UsersController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class UsersController extends Controller
+{
+    // 用户注册操作
+    public function create()
+    {
+        return '用户注册表单';
+    }
+}
