+++ routes/web.php
@@ -11,6 +11,11 @@
 |
 */
 
-Route::get('/', function () {
-    return view('welcome');
-});
+/*静态页面 例如header footer, 静态Body*/
+Route::get('/', 'StaticPagesController@home')->name('home');
+
+/*用户相关操作 注册， 修改密码， 用户激活等*/
+Route::get('signup','UsersController@create')->name('signup');
+
+/*用户状态操作 登录， 退出等*/
+Route::get('login', 'SessionsController@create')->name('login');
